правки после проверки + отдельный модуль для комментариев
